recrodsの型定義
